requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
psycopg2-binary>=2.9
alembic>=1.14
pydantic-settings>=2.7
google-genai>=1.0
httpx>=0.28
celery[redis]>=5.4
python-dotenv>=1.0
